chore: drop reconcile system and active-job telemetry
The `_reconcile_interrupted` classifier and its sub-handlers were
already dead in production (only tests called them). The legacy PR
label migration is also obsolete — publish.py deletes the old
labels and no open PR still carries them.

Changes:
- Rename `cai_lib/cmd_lifecycle.py` → `cai_lib/watchdog.py`, keeping
  only `_rollback_stale_in_progress` (the stale-lock watchdog).
- Delete `_reconcile_interrupted`, `_reconcile_fix/revise/refine`,
  `_migrate_legacy_pr_pipeline_labels`, and `_LEGACY_PR_LABEL_MAP`.
- Remove the `_write_active_job` / `_clear_active_job` observability
  telemetry (40+ call sites in cai.py, both helper functions, the
  `ACTIVE_JOB_PATH` config constant, and `cai-active.json` docs).
- Drop `tests/test_reconcile.py`; retarget `tests/test_rollback.py`
  patch paths to `cai_lib.watchdog`.
- Refresh `CODEBASE_INDEX.md` and `scripts/generate-index.sh` for
  the renamed module and removed test.

FSM labels are now the single source of truth for lifecycle state;
the next step (follow-up PR) is to make each cmd_* driver
idempotent on pickup so the stale-lock watchdog can eventually be
simplified too.
