+++ cai_lib/__init__.py
@@ -26,7 +26,6 @@ from cai_lib.config import (
     COST_LOG_PATH,
     REVIEW_PR_PATTERN_LOG,
     OUTCOME_LOG_PATH,
-    ACTIVE_JOB_PATH,
     LABEL_RAISED,
     LABEL_IN_PROGRESS,
     LABEL_PR_OPEN,
@@ -57,8 +56,6 @@ from cai_lib.config import (
 from cai_lib.logging_utils import (
     log_run,
     log_cost,
-    _write_active_job,
-    _clear_active_job,
     _get_issue_category,
     _log_outcome,
     _load_outcome_counts,
@@ -81,7 +78,7 @@ from cai_lib.github import (
     _build_implement_user_message,
 )
 
-from cai_lib.cmd_lifecycle import _rollback_stale_in_progress, _reconcile_interrupted
+from cai_lib.watchdog import _rollback_stale_in_progress
 
 from cai_lib.cmd_implement import _parse_decomposition
 
@@ -97,7 +94,7 @@ __all__ = [
     "CODE_AUDIT_MEMORY", "PROPOSE_MEMORY", "UPDATE_CHECK_MEMORY",
     "COST_OPTIMIZE_MEMORY", "AGENT_MEMORY_DIR",
     "LOG_PATH", "COST_LOG_PATH", "REVIEW_PR_PATTERN_LOG",
-    "OUTCOME_LOG_PATH", "ACTIVE_JOB_PATH",
+    "OUTCOME_LOG_PATH",
     "LABEL_RAISED", "LABEL_IN_PROGRESS", "LABEL_PR_OPEN",
     "LABEL_MERGED", "LABEL_SOLVED", "LABEL_NO_ACTION", "LABEL_NEEDS_SPIKE",
     "LABEL_NEEDS_EXPLORATION", "LABEL_REFINED", "LABEL_REVISING", "LABEL_PARENT",
@@ -108,7 +105,7 @@ __all__ = [
     "_STALE_IN_PROGRESS_HOURS", "_STALE_REVISING_HOURS",
     "_STALE_NO_ACTION_DAYS", "_STALE_MERGED_DAYS",
     # logging
-    "log_run", "log_cost", "_write_active_job", "_clear_active_job",
+    "log_run", "log_cost",
     "_get_issue_category", "_log_outcome", "_load_outcome_counts",
     "_load_outcome_stats", "_load_cost_log", "_row_ts", "_build_cost_summary",
     # subprocess
@@ -116,9 +113,8 @@ __all__ = [
     # github
     "_gh_json", "check_gh_auth", "check_claude_auth", "_transcript_dir_is_empty",
     "_set_labels", "_issue_has_label", "_build_issue_block", "_build_implement_user_message",
-    # cmd_lifecycle
+    # watchdog
     "_rollback_stale_in_progress",
-    "_reconcile_interrupted",
     # cmd_implement
     "_parse_decomposition",
     # fsm
